Separate out Brightway functionality
